mempy/requirements.txt
Flask
pymongo
azure-ai-textanalytics